requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
Django==3.0.5
django-environ==0.4.5
django-user-agents==0.4.0
idna==2.10
lxml==4.6.1
psycopg2-binary==2.8.6
python-memcached==1.59
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tmdbsimple==2.6.6
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
